chore(repo): pull restore out of prepare script
We use different versions of dotnet in our smoke tests, which was causing failures for the .NET 6 matricies as our demo project targets .NET 7. This pulls `dotnet restore` out of the smoke test script. Since it uses the published versions of nx and our plugins it does not need to build our packages. For workflows that do need to build our packages, restore is ran automatically as part of the build.